Parse any `BinaryFloatingPoint` type
